LayoutTests/imported/w3c/web-platform-tests/selection/fire-selectionchange-event-on-pressing-backspace-expected.txt

wld


FAIL Selectionchange event is fired after removing the character assert_equals: target.innerHTML should match expected "<p>wo<i id=\"caret\">l</i>d</p>" but got "<p>word</p>"
PASS Selectionchange event is fired after removing the range
